Correlate and preserve mirrored event records
