+++ vibeguard-runtime/src/lib.rs
@@ -4,6 +4,7 @@ mod event_schema;
 mod hook_checks {
     pub mod bash;
     pub mod common;
+    pub mod jsonl;
 }
 mod hook_input_diag;
 mod pkg_rewrite;
